api/pyproject.toml
[project]
name = "langperf-api"
version = "0.1.0"
description = "LangPerf backend — OTLP ingestion and UI API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.30",
    "opentelemetry-proto>=1.28",
    "protobuf>=4.25",
    "python-multipart>=0.0.12",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
